Add PlaceholderManager
This class allows to register placeholders which can be used to replace
values in strings.

The placeholders are stored in a dictionary and can be accessed by the
key of the placeholder.

The placeholder manager also provides a method to resolve a string
by replacing all placeholders with their values. This is useful for
displaying dynamic text in the UI.